First Beta release.

+++ trunk/imsdroid/src/org/doubango/imsdroid/utils/UriUtils.java
@@ -31,22 +31,24 @@ import org.doubango.tinyWRAP.SipUri;
 public class UriUtils {
 
 	private final static long MAX_PHONE_NUMBER = 1000000000000L;
-	private final static String INVALID_SIP_URI = "sip:invalid@open-ims.tes";
+	private final static String INVALID_SIP_URI = "sip:invalid@open-ims.test";
 	
 	public static String getDisplayName(String uri){
 		String displayname = null;
-		Group.Contact contact = ServiceManager.getContactService().getContact(uri);
-		if(contact != null  && (displayname = contact.getDisplayName()) != null){
-			return displayname;
+		if(!StringUtils.isNullOrEmpty(uri)){
+			Group.Contact contact = ServiceManager.getContactService().getContact(uri);
+			if(contact != null  && (displayname = contact.getDisplayName()) != null){
+				return displayname;
+			}
+			
+			SipUri sipUri = new SipUri(uri);
+			if(sipUri.isValid()){
+				displayname = sipUri.getUserName();
+			}
+			sipUri.delete();
 		}
 		
-		SipUri sipUri = new SipUri(uri);
-		if(sipUri.isValid()){
-			displayname = sipUri.getUserName();
-		}
-		sipUri.delete();
-		
-		return displayname;
+		return displayname == null ? uri : displayname;
 	}
 	
 	public static boolean isValidSipUri(String uri){
@@ -77,26 +79,40 @@ public class UriUtils {
 	}
 	
 	public static String getValidPhoneNumber(String uri){
-		SipUri sipUri = new SipUri(uri);
-		if(sipUri.isValid()){
-			String userName = sipUri.getUserName();
-			if(userName != null){
-				try{
-					String scheme = sipUri.getScheme();
-					if(scheme != null && scheme.equals("tel")){
-						userName = userName.replace("-", "");
+		if(uri != null && (uri.startsWith("sip:") || uri.startsWith("sip:") || uri.startsWith("tel:"))){
+			SipUri sipUri = new SipUri(uri);
+			if(sipUri.isValid()){
+				String userName = sipUri.getUserName();
+				if(userName != null){
+					try{
+						String scheme = sipUri.getScheme();
+						if(scheme != null && scheme.equals("tel")){
+							userName = userName.replace("-", "");
+						}
+						long result = Long.parseLong(userName.startsWith("+") ? userName.substring(1) : userName);
+						if(result < UriUtils.MAX_PHONE_NUMBER){
+							return userName;
+						}
 					}
-					long result = Long.parseLong(userName.startsWith("+") ? userName.substring(1) : userName);
-					if(result < UriUtils.MAX_PHONE_NUMBER){
-						return userName;
+					catch(NumberFormatException ne){ }
+					catch (Exception e){
+						e.printStackTrace();
 					}
 				}
-				catch (Exception e){
-					e.printStackTrace();
+			}
+			sipUri.delete();
+		}
+		else{
+			try{
+				uri = uri.replace("-", "");
+				long result = Long.parseLong(uri.startsWith("+") ? uri.substring(1) : uri);
+				if(result < UriUtils.MAX_PHONE_NUMBER){
+					return uri;
 				}
-				finally{
-					sipUri.delete();
-				}
+			}
+			catch(NumberFormatException ne){ }
+			catch (Exception e) {
+				e.printStackTrace();
 			}
 		}
 		return null;
